+++ tests/test_files/type_alias.py
@@ -1,0 +1,2 @@
+type Integer = int
+type ComplexAlias[T: (int and str and bytes)] = list[T] | set[T] | dict[T, T]